frontend/public/assets/apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose==3.3.0
psycopg[binary]==3.1.19
aiosqlite==0.20.0
pypdf==4.2.0
reportlab==4.1.0
